Workaround probable Windows bug in path equivalency checks
In Windows, std::filesystem::equivalent throws
std::system_error if a path contains characters that can't be
represented in the system code page (e.g. Windows 1252). This is
undocumented and seems unnecessary as Microsoft's own docs say Windows
paths are stored natively in Unicode, so it sounds like a bug to me.

To work around this, return true early if the path objects are equal,
so std::filesystem::equivalent doesn't need to be called, and catch
system_error if it does need to be called - this may result in false
negatives, but for LOOT's purposes that is fine and is very unlikely
anyway.