Don't include SSH support when building libgit2
Windows CI builds are currently failing with the following error:

  -- Checking for module 'libssh2'
  --   Found libssh2, version 1.8.2
  CMake Error at cmake/FindPkgLibraries.cmake:17 (MESSAGE):
    could not resolve ssh2
  Call Stack (most recent call first):
    src/CMakeLists.txt:213 (FIND_PKGLIBRARIES)

this is odd, but can be avoided by not building in SSH support. LOOT's
masterlists are all downloaded over HTTPS, so this wouldn't break
anything (unless there are people hosting their own masterlists that
aren't accessible over HTTPS, but that's very unlikely).