Upload CI build artifacts to GitHub Actions
Bintray will be shut down in May, so we need to find another artifact
store before then.

GitHub Actions is one option, and allows builds to store artifacts
for up to 90 days, with apparently no storage size limit for public
repositories. 90 days is good enough to temporarily host builds for
loot/loot to use for integration testing before release.

The downside is that you need to be logged into a GitHub account in
order to download a build artifact uploaded to GitHub Actions.