Compare language codes in MessageContent::Choose()
If there is no exact match, compare language codes so that a more
generic match is preferred over just returning the English content.

This means that if a locale code is given, Choose() will also look for
the locale's language code (but will not match another locale with the
same language code), and if a language code is given, Choose() will also
look for a locale with that language code.