Update CI to build using MSVC 2019
GitHub are going to be removing their windows-2016 images in March 2022,
and there's a brownout today. I've been using MSVC 2019 locally for a
while (a year?) without any problems, and LOOT itself is now using MSVC
2019.

This means that anything linking to the binaries needs to use MSVC 2019
or newer. The details are here:
<https://docs.microsoft.com/en-us/cpp/porting/binary-compat-2015-2017?view=msvc-170>.