Overhaul how group edges are added during sorting
The new logic is conceptually simpler, with fewer special cases to
handle. Unlike the old implementation, the new approach avoids cycles.

It basically does depth-first searches through the group graph, adding
edges between each group's plugins and the plugins in the group's
successors.

DFSes are run in the order that groups are defined, except that root
vertices go first. That exception is so that in the common case of a
single linear collection of groups, the order those groups are defined
in doesn't matter. The order still matters if there are multiple root
nodes or the path branches.

A more efficient solution could be to implement a custom DFS algorithm
that doesn't stop when it reaches a vertex it's already visited (which
would be fine since the graph has already been validated to be acyclic),
as then only the root vertices would need to be searched from.

The order in which load after groups are listed doesn't matter, because
the DFSes run in the opposite direction.

I'm pretty sure that the order that plugins within groups are iterated
over doesn't matter, I tested it manually with my test load order and
saw no difference when the order was randomly shuffled. I've added a
test that tries to prove that vertex order doesn't matter but there may
well be more complex scenarios for which order could matter but which
wouldn't be caught by this test.

A buffer is used to hold the plugins in the previous groups in the
current path, because that's noticeably faster than just recording the
path and looking up the plugins for each group in the path. It does
duplicate the group vectors, but that's an insignificant amount of
memory used.

This involved reversing the direction of edges in the group graph, as
this switches the logic from trying to find predecessors to trying to
successors: the new direction matches that of plugin graph edges, so
it's less confusing.

While one of my goals for doing this was to reduce the number of edges
added and so increase sorting, in my test load order the number of group
edges increased slightly, the number of overall edges decreased
slightly, and sorting performance was unchanged. The resulting load
order is identical.

My initial reimplementation only ran DFSes from root vertices and only
buffered plugins to be carried forwards if any of their edges were
skipped, but it couldn't handle a graph with paths that merged or some
cases where an ancestor plugin should get an edge to a successor even
when the current group's plugins can't.