Make the CMake config for tests optional
Managed using the new LIBLOOT_BUILD_TESTS config option, which defaults to ON.

Also manage the Google Test and testing-plugins dependencies using FetchContent, as it integrates more neatly with FindPackage.

This means that if you're not interested in the tests, you don't need to build them or download and build Google Test or download the testing plugins.