Don't throw during sort if a non-blueprint-master has a blueprint master as a master
The game will not load the blueprint master before the other plugin, so if the latter overrides any records from the former, those changes won't be seen in-game.

There are probably many (pre-CK) plugins that have the official BlueprintShips-Starfield.esm blueprint master as a master but not actually override any of its records. Ideally such plugins would be updated to remove the unnecessary master or uninstalled, but as having it doesn't seem to do any harm, I don't think it's worth preventing LOOT from sorting in this situation.