Remove ability to use find_package for yaml-cpp dependency
If the build system happens to have yaml-cpp installed as a system package it would link against that, causing non-obvious errors when libloot is then unable to resolve masterlist merge keys. As such, make it so that can't accidentally happen.

Also document an alternative approach for getting FetchContent to use existing sources instead of trying to download them, which is relevant for building LOOT's Flatpak.