Add support for emitting YAML with anchors and aliases
Support is currently limited to aliasing:

- Condition and constraint strings
- File values
- Message values
- MessageContent arrays (including single-value arrays that are
  serialised as strings)

An anchor is written if the same value appears more than once in the
metadata document being written, and if an anchor for that value has not
already been written. If a value has already been written with an
anchor, later appearances of that value will be written as aliases of
that anchor.

Anchors are named according to the type of data they're for, followed by
an incrementing number, e.g. file1, message1, contents1, condition1.

This changes the order in which plugins are written to the masterlist,
so that it's consistent.